[Kernel] Remove Parquet utility methods into a separate utils file
