+++ kernel/kernel-defaults/src/main/java/io/delta/kernel/defaults/internal/parquet/RowConverter.java
@@ -33,7 +33,7 @@ import static io.delta.kernel.internal.util.Preconditions.checkArgument;
 
 import io.delta.kernel.defaults.internal.data.DefaultColumnarBatch;
 import io.delta.kernel.defaults.internal.data.vector.DefaultStructVector;
-import static io.delta.kernel.defaults.internal.DefaultKernelUtils.findSubFieldType;
+import static io.delta.kernel.defaults.internal.parquet.ParquetSchemaUtils.findSubFieldType;
 
 class RowConverter
     extends GroupConverter
